Minimize net diff of binder-gen visibility/identifier-clash fix
